requirements.txt
streamlit>=1.32.0
groq>=0.5.0
python-dotenv>=1.0.0
gitpython>=3.1.40
faiss-cpu>=1.7.4
sentence-transformers>=2.6.1
tiktoken>=0.6.0
pygments>=2.17.0
markdown>=3.5.2
beautifulsoup4>=4.12.3
chardet>=5.2.0
numpy>=1.24.0